+++ CHANGELOG.rst
@@ -6,7 +6,8 @@ Changelog
 
 Features:
 
-* Expose a public structured marker tree via ``Marker.as_ast()`` (:pull:`1145`)
+* Expose a public structured marker tree via ``Marker.as_ast()`` in (:pull:`1145`)
+* Add ``Marker.from_ast()`` to construct a ``Marker`` from a ``MarkerNode`` tree for marker construction and mutation in (:pull:`1145`)
 
 26.2 - 2026-04-24
 ~~~~~~~~~~~~~~~~~